refactor: Clean up import statements and update TypeScript configuration
