+++ frontend/src/components/LandingPage/About.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const About: React.FC = () => {
   return (
     <>
