+++ frontend/src/components/LandingPage/Banner.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 const Banner: React.FC = () => {
   return (
     <>
