+++ frontend/src/components/LandingPage/BetaForm.tsx
@@ -1,5 +1,5 @@
-import { useState } from 'react';
 import axios, { AxiosError } from 'axios';
+import { useState } from 'react';
 import { useForm } from 'react-hook-form';
 
 interface FormData {
