+++ frontend/src/hooks/useSearchToursAPI.ts
@@ -1,6 +1,6 @@
+import type { SearchResponse } from '@/api/api';
+import { searchToursAPI } from '@/api/api';
 import { useQuery, UseQueryResult } from '@tanstack/react-query';
-import { searchToursAPI } from '@/api/api.js';
-import type { SearchResponse } from '@/api/api.js';
 
 export default function useSearchToursAPI(url: string): UseQueryResult<SearchResponse, Error> {
   return useQuery({
