+++ frontend/src/main.tsx
@@ -1,8 +1,8 @@
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import App from './App';
 import './assets/all.scss';
-import App from './App.tsx';
 
 const queryClient = new QueryClient();
